Carry over descriptions (#142)
* Updates to element short and definition

Support short definitions derived from element descriptions.  Use the element description as short/definition when element has a type constraint.

We don't want to *always* use the element description because that can make lots of noise in the differential for very little value.

* Fix algorithm for deriving short from description

Properly handle edge cases for `.` when it does not indicate the end of a sentence.

* Add short descriptions to short for slices

* Add comments for truncation code