v6.1.0: Element type enforcement (#99)
* Remove export step from unit tests

Now that we are on CIMPL6, it does not make sense to test the CIMPL6 exporter anymore.  Removing this simplifies the test framework.

Due to causing changes in indentation, however, it essentially affects every single line in the import-test.js file.

* Type Rules: Like Inherits from Like + No Value Redeclaration

Enforce rules such that:
- Entries should only inherit from Entries or Abstracts
- Abstracts should only inherit from Abstracts
- Elements should only inherit from Elements
- Groups should only inherit from Groups
- An element cannot redeclare Value if it has a parent

In these cases, we report an ERROR, but we do proceed as best we can with what the author is attempting to do.  This is backwards compatible with the current approach.

* Enforce more rules around Elements and non-Elements

1. Non-elements cannot declare Value:
2. Non-elements cannot constrain root Value (does not exist)
2. Elements cannot declare Property:
3. Elements cannot constrain root properties (do not exist)

* Assign error numbers to type enforcement errors

* Temporarily point shr-models to element-type-enforcement branch

* v6.1.0